requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy==3.1.1
requests==2.32.5
openai==1.107.2
python-dotenv==1.0.0
beautifulsoup4==4.13.5
gunicorn==21.2.0
